requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fonttools==4.58.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
motor==3.7.1
numpy==2.2.6
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.33.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
threadpoolctl==3.6.0
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0